+++ google-beta/cloudfunctions_operation.go
@@ -3,41 +3,21 @@
 package google
 
 import (
-	"fmt"
 	"time"
 
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	tpgcloudfunctions "github.com/hashicorp/terraform-provider-google-beta/google-beta/services/cloudfunctions"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 	"google.golang.org/api/cloudfunctions/v1"
 )
 
-type CloudFunctionsOperationWaiter struct {
-	Service *cloudfunctions.Service
-	tpgresource.CommonOperationWaiter
-}
-
-func (w *CloudFunctionsOperationWaiter) QueryOp() (interface{}, error) {
-	if w == nil {
-		return nil, fmt.Errorf("Cannot query operation, it's unset or nil.")
-	}
-	return w.Service.Operations.Get(w.Op.Name).Do()
-}
-
+// Deprecated: For backward compatibility cloudFunctionsOperationWait is still working,
+// but all new code should use CloudFunctionsOperationWait in the tpgcloudfunctions package instead.
 func cloudFunctionsOperationWait(config *transport_tpg.Config, op *cloudfunctions.Operation, activity, userAgent string, timeout time.Duration) error {
-	w := &CloudFunctionsOperationWaiter{
-		Service: config.NewCloudFunctionsClient(userAgent),
-	}
-	if err := w.SetOp(op); err != nil {
-		return err
-	}
-	return tpgresource.OperationWait(w, activity, timeout, config.PollInterval)
+	return tpgcloudfunctions.CloudFunctionsOperationWait(config, op, activity, userAgent, timeout)
 }
 
+// Deprecated: For backward compatibility IsCloudFunctionsSourceCodeError is still working,
+// but all new code should use IsCloudFunctionsSourceCodeError in the tpgcloudfunctions package instead.
 func IsCloudFunctionsSourceCodeError(err error) (bool, string) {
-	if operr, ok := err.(*tpgresource.CommonOpError); ok {
-		if operr.Code == 3 && operr.Message == "Failed to retrieve function source code" {
-			return true, fmt.Sprintf("Retry on Function failing to pull code from GCS")
-		}
-	}
-	return false, ""
+	return tpgcloudfunctions.IsCloudFunctionsSourceCodeError(err)
 }
