+++ google-beta/service_networking_operation.go
@@ -5,35 +5,13 @@ package google
 import (
 	"time"
 
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	tpgservicenetworking "github.com/hashicorp/terraform-provider-google-beta/google-beta/services/servicenetworking"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 	"google.golang.org/api/servicenetworking/v1"
 )
 
-type ServiceNetworkingOperationWaiter struct {
-	Service             *servicenetworking.APIService
-	Project             string
-	UserProjectOverride bool
-	tpgresource.CommonOperationWaiter
-}
-
-func (w *ServiceNetworkingOperationWaiter) QueryOp() (interface{}, error) {
-	opGetCall := w.Service.Operations.Get(w.Op.Name)
-	if w.UserProjectOverride {
-		opGetCall.Header().Add("X-Goog-User-Project", w.Project)
-	}
-	return opGetCall.Do()
-}
-
+// Deprecated: For backward compatibility ServiceNetworkingOperationWaitTime is still working,
+// but all new code should use ServiceNetworkingOperationWaitTime in the tpgservicenetworking package instead.
 func ServiceNetworkingOperationWaitTime(config *transport_tpg.Config, op *servicenetworking.Operation, activity, userAgent, project string, timeout time.Duration) error {
-	w := &ServiceNetworkingOperationWaiter{
-		Service:             config.NewServiceNetworkingClient(userAgent),
-		Project:             project,
-		UserProjectOverride: config.UserProjectOverride,
-	}
-
-	if err := w.SetOp(op); err != nil {
-		return err
-	}
-	return tpgresource.OperationWait(w, activity, timeout, config.PollInterval)
+	return tpgservicenetworking.ServiceNetworkingOperationWaitTime(config, op, activity, userAgent, project, timeout)
 }
