+++ google-beta/serviceman_operation.go
@@ -3,38 +3,16 @@
 package google
 
 import (
-	"fmt"
 	"time"
 
-	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	tpgservicemanagement "github.com/hashicorp/terraform-provider-google-beta/google-beta/services/servicemanagement"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 	"google.golang.org/api/googleapi"
 	"google.golang.org/api/servicemanagement/v1"
 )
 
-type ServiceManagementOperationWaiter struct {
-	Service *servicemanagement.APIService
-	tpgresource.CommonOperationWaiter
-}
-
-func (w *ServiceManagementOperationWaiter) QueryOp() (interface{}, error) {
-	if w == nil {
-		return nil, fmt.Errorf("Cannot query operation, it's unset or nil.")
-	}
-	return w.Service.Operations.Get(w.Op.Name).Do()
-}
-
+// Deprecated: For backward compatibility ServiceManagementOperationWaitTime is still working,
+// but all new code should use ServiceManagementOperationWaitTime in the tpgservicemanagement package instead.
 func ServiceManagementOperationWaitTime(config *transport_tpg.Config, op *servicemanagement.Operation, activity, userAgent string, timeout time.Duration) (googleapi.RawMessage, error) {
-	w := &ServiceManagementOperationWaiter{
-		Service: config.NewServiceManClient(userAgent),
-	}
-
-	if err := w.SetOp(op); err != nil {
-		return nil, err
-	}
-
-	if err := tpgresource.OperationWait(w, activity, timeout, config.PollInterval); err != nil {
-		return nil, err
-	}
-	return w.Op.Response, nil
+	return tpgservicemanagement.ServiceManagementOperationWaitTime(config, op, activity, userAgent, timeout)
 }
