+++ google-beta/services/compute/compute_operation_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package compute
 
 import (
 	"fmt"
